Add icon and label selection
