+++ app/lib/selections/selection.dart
@@ -3,6 +3,7 @@ import 'package:butterfly/models/property.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 import '../models/element.dart';
 import '../models/painter.dart';
@@ -11,8 +12,11 @@ import '../widgets/color_field.dart';
 import '../widgets/exact_slider.dart';
 
 part 'elements/element.dart';
+
 part 'painters/painter.dart';
 part 'painters/pen.dart';
+part 'painters/label.dart';
+
 part 'hand.dart';
 
 abstract class Selection<T> {
@@ -34,6 +38,7 @@ abstract class Selection<T> {
   }
 
   String getLocalizedName(BuildContext context);
+  IconData getIcon({bool filled = false});
 
   List<Widget> buildProperties(BuildContext context) => [];
